src/react-native-explorer/agent-server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
mcp>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
websockets>=12.0
pyyaml>=6.0
